Move util.lua to the specs directory
Now that this repo no longer has a Lua module, util.lua is the last Lua file
remaining in the src directory. Let's move it to the spec directory, and leave
only C stuff in the src.
